# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoreBanking API
  version: 1.0.0-phase1
  description: |
    Staff and integration API for CoreBanking. Every request is tenant-scoped: the tenant comes from the
    verified access token (`tenant` claim), never from the URL or body.

    Conventions
    - Money is a decimal **string** (e.g. "100000.00"). Never a JSON number.
    - Dates are ISO `YYYY-MM-DD` business dates; timestamps are RFC 3339 UTC.
    - Changes to master data and money go through maker-checker: the call returns **202** with an
      approval request; the change is applied when a different user approves it.
    - Errors use RFC 9457 `application/problem+json`.
    - `Idempotency-Key` header is honoured on all POSTs that create things.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearer: []
...
paths:
...
  /api/v1/loan-products:
    get:
      tags: [Lending]
      operationId: listLoanProducts
      summary: Loan products with fee rules
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/LoanProduct' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Lending]
      operationId: proposeLoanProduct
      summary: Create or change a product (maker-checker; approval bumps the version)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoanProduct' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loan-products/{code}:
    get:
      tags: [Lending]
      operationId: getLoanProduct
      summary: One loan product
      parameters: [ { name: code, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanProduct' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loan-product-templates:
    get:
      tags: [Lending]
      operationId: listLoanProductTemplates
      summary: Product templates - starting points for the product wizard (US-038)
      description: >-
        Seeded examples (personal EMI, business EMI and step-up, gold bullet, weekly micro-loan, flat-rate consumer
        loan, floating home loan in tranches, seasonal structured loan). `product` has the shape of LoanProduct
        without code, status and version; the figures are illustrative and nothing here is a live product.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/LoanProductTemplate' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loan-products/preview:
    post:
      tags: [Lending]
      operationId: previewLoanProduct
      summary: Simulate an account on a draft product - schedule, fees, net disbursal, APR (US-044; nothing is stored)
      description: >-
        The product need not be saved or approved, and needs no code. The sample loan defaults to the product's
        smallest amount, shortest tenor and lowest rate. GST is shown as an intra-state supply. For a STRUCTURED
        product without `scheduleRows`, equal principal on each period date is shown as an example
        (`sampleSchedule: true`). The engine is the one that books loans, so a product the engine would refuse is
        refused here (422).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductPreviewRequest' }
      responses:
        '200':
          description: KFS figures of the sample loan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanKfs' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/preview:
    post:
      tags: [Lending]
      operationId: previewLoan
      summary: Schedule, fees, net disbursal and APR from the same engine that posts (KFS figures)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoanApplication' }
      responses:
        '200':
          description: KFS figures
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanKfs' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans:
    get:
      tags: [Lending]
      operationId: searchLoans
      summary: Search loans by number, customer, name or LOS reference
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/LoanStatus' } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/LoanSummary' } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Lending]
      operationId: createLoan
      summary: Book a loan as SANCTIONED (idempotent on externalRef); money moves at disbursement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoanApplication' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/loans/{id}/schedule:
    get:
      tags: [Lending]
      operationId: getLoanSchedule
      summary: Demands raised, charges and the future schedule
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanSchedule' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/transactions:
    get:
      tags: [Lending]
      operationId: getLoanTransactions
      summary: Financial transactions (newest first); day-end accruals are not listed
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/LoanTxn' } }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/parties:
    get:
      tags: [Lending]
      operationId: listLoanParties
      summary: Borrower, co-applicants and guarantors of the loan
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: The borrower first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LoanParty' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/kfs:
    get:
      tags: [Lending]
      operationId: getLoanKfs
      summary: Key Fact Statement as generated at booking
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanKfs' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/kfs-acceptance:
    post:
      tags: [Lending]
      operationId: acceptLoanKfs
      summary: Record the borrower's KFS acceptance (required before disbursement)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channel: { type: string, examples: [OTP, ESIGN, BRANCH] }
                evidenceRef: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/disbursement:
    post:
      tags: [Lending]
      operationId: disburseLoan
      summary: Staff users get a 202 approval; LOS clients with loan:stp are disbursed immediately (200)
      description: >-
        `amount` is what to disburse now; omitted, it is everything not yet drawn. A part of the sanctioned amount
        can be drawn only on a product with `multipleDisbursements`; further tranches use this same endpoint while
        the loan is ACTIVE with an undrawn amount. Interest accrues only on what is drawn. DISBURSEMENT fees are
        charged once, on the first tranche, on the sanctioned amount; EVERY_DISBURSEMENT fees on each tranche's
        amount; broken-period interest in mode DEDUCT_AT_DISBURSAL is taken from the first tranche. The schedule is
        rebuilt on each tranche: with `preEmi` the instalments are interest-only until the loan is fully drawn (or
        the undrawn amount is cancelled) and the EMIs then run for the full tenor; otherwise the instalments are
        recomputed on the amount drawn over the instalments left. No tranche is paid while the account has unpaid
        dues or is NPA (409). A tranche cannot be reversed.
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: [number, string], description: 'Amount of this disbursement; default: all that is undrawn' }
                beneficiaryName: { type: string }
                beneficiaryAccount: { type: string }
                ifsc: { type: string }
                mode: { type: string, examples: [IMPS, NEFT, CASH] }
      responses:
        '200':
          description: Disbursed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/tranches:
    get:
      tags: [Lending]
      operationId: listLoanTranches
      summary: Disbursements made so far and the amount still undrawn (US-050)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoanTranches' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/simulations/disbursement:
    post:
      tags: [Lending]
      operationId: simulateLoanDisbursement
      summary: What a disbursement would do today - fees, interest deducted, net payout, schedule (nothing changes)
      description: >-
        For a SANCTIONED loan the first disbursement, for an ACTIVE loan the next tranche. Computed by the code that
        posts the disbursement; nothing is posted or stored. The same refusals apply (409 / 422).
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: [number, string, 'null'], description: 'Default: all that is undrawn' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DisbursementSimulation' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/simulations/transaction:
    post:
      tags: [Lending]
      operationId: simulateLoanTransaction
      summary: What a receipt, part-prepayment or pre-closure would do on a date (nothing changes)
      description: >-
        `onDate` is today or up to 366 days ahead: the day-ends until then (demands, accrual, penal charges,
        classification) are run on a copy first. The figures come from the engine code that posts the transaction;
        nothing is posted or stored. Restructure options are simulated by `/restructure/simulation`.
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionSimulationRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionSimulation' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/sanction-change/preview:
    post:
      tags: [Lending]
      operationId: previewLoanSanctionChange
      summary: Before/after of a change of the sanctioned amount (nothing changes)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SanctionChangeRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SanctionChangePreview' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/sanction-change:
    post:
      tags: [Lending]
      operationId: proposeLoanSanctionChange
      summary: Propose a top-up, a reduction or the cancellation of the undrawn amount (maker-checker, entity LOAN_SANCTION_CHANGE)
      description: >-
        **Reduction** may only take away undrawn amount (never below what is disbursed); `cancelUndrawn` brings the
        sanctioned amount down to the amount disbursed, which starts the EMIs of a pre-EMI loan. **Top-up** in the
        same account needs a product with `topUpAllowed`, an amount within the product maximum and the customer's
        exposure limit, and a STANDARD account with no unpaid dues that is not under post-restructuring monitoring
        (409 otherwise: new money to a borrower in arrears would be evergreening). No money moves on approval: the
        extra amount is then paid out through `/disbursement`. One checker. Recorded in the amendment history
        (kind SANCTION_CHANGE); it cannot be reversed. Needs `loan:amend`.
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SanctionChangeRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/npa-override:
    post:
      tags: [Lending]
      operationId: proposeLoanNpaOverride
      summary: Manual NPA mark - downgrade the account to an NPA class or hold it there (maker-checker, two checkers)
      description: >-
        Entity LOAN_NPA_OVERRIDE, action OVERRIDE. The override sets the worst class the account may be upgraded to
        until `until`: it can downgrade the account or hold it at its present NPA class, never upgrade it (409).
        A downgrade from a performing class moves unrealised interest to suspense. While it holds, day-end
        classification can make the class worse but not better; after it expires the normal rules apply. LOSS is
        permanent. Recorded in the amendment history (kind NPA_OVERRIDE); it cannot be reversed. Needs `loan:classify`.
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NpaOverrideRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/npa-override/release:
    post:
      tags: [Lending]
      operationId: proposeLoanNpaRelease
      summary: Un-mark - end a manual NPA override before its expiry (maker-checker, two checkers)
      description: >-
        Entity LOAN_NPA_OVERRIDE, action RELEASE. Refused (409) while the account has unpaid dues, when proposed
        and again when approved: an NPA is upgraded only when all arrears are cleared (RBI IRACP norms). Nothing is
        posted; the next day-end classifies the account by the normal rules. A LOSS classification cannot be
        released. Needs `loan:classify`.
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
...
  /api/v1/loans/{id}/prepayments:
    post:
      tags: [Lending]
      operationId: prepayLoan
      summary: Part-prepayment; the schedule is rebuilt (reduce EMI or reduce tenure)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { $ref: '#/components/schemas/Money' }
                mode: { type: string, enum: [REDUCE_EMI, REDUCE_TENURE] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/preclosure-quote:
    get:
      tags: [Lending]
      operationId: getPreclosureQuote
      summary: Pre-closure amount today (the posting uses the same figures)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PreclosureQuote' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/preclosure:
    post:
      tags: [Lending]
      operationId: precloseLoan
      summary: Close the loan; amount must equal the quote total
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AmountBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/cancellation-quote:
    get:
      tags: [Lending]
      operationId: getCancellationQuote
      summary: Amount to exit within the cooling-off period
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { $ref: '#/components/schemas/Money' }
                  asOf: { type: string, format: date }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/cancellation:
    post:
      tags: [Lending]
      operationId: cancelLoan
      summary: Cooling-off exit (principal + interest for days used; disclosed fees retained)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AmountBody' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/charges:
    post:
      tags: [Lending]
      operationId: chargeLoanFee
      summary: Charge a product fee (with GST) to the loan
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [feeCode]
              properties:
                feeCode: { type: string }
                base: { $ref: '#/components/schemas/Money' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Loan' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/charges/{chargeId}/waiver:
    post:
      tags: [Lending]
      operationId: waiveLoanCharge
      summary: Propose a waiver of an unpaid charge (maker-checker)
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: chargeId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, reason]
              properties:
                amount: { $ref: '#/components/schemas/Money' }
                reason: { type: string }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/loans/{id}/transactions/{txnId}/reverse:
    post:
      tags: [Lending]
      operationId: reverseLoanTransaction
      summary: Propose reversal of a transaction and every later one; days since are replayed (maker-checker)
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { name: txnId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonBody' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Accepted:
      description: Change recorded as an approval request; applied when a checker approves it
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Approval' }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    LoanStatus:
      type: string
      enum: [SANCTIONED, ACTIVE, FROZEN, CLOSED, CANCELLED, WRITTEN_OFF]
...
    LoanProductTemplate:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string, enum: [RETAIL, BUSINESS, GOLD, MICRO, CONSUMER, HOUSING, AGRI] }
        product: { type: object, additionalProperties: true, description: 'LoanProduct without code, status and version' }
...